+++ src/components/CategoryShortDescription.vue
@@ -2,7 +2,7 @@
   <template v-if="!!html">
     <div :class="`propeller-category-short-description mb-6 ${className || ''}`">
       <div
-        class="propeller-category-short-description__content prose prose-slate max-w-none text-muted-foreground"
+        class="propeller-category-short-description__content text-muted-foreground"
         v-html="html"
       ></div>
     </div>